transform double to integer birthweight

+++ src/calvingDataMap.h
@@ -44,7 +44,7 @@ public:
   string getVerifiedTvdNr(string id);
   string verifyBreed(string breedstr, string indstr);
   string verifySexBirth(string sexstr, string indstr);
-  double verifyBirthWeight(string birthweightstr, string indstr);
+  int verifyBirthWeight(string birthweightstr, string indstr);
   string verifyMultiple(string multiplesstr, string indstr);
   int verifyAbort(string abortstr, string indstr);
   int verifyCalvingscore(string calvingscorestr, string indstr);
